Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is there a chance of ct_state becoming -trk from ct in any other scenarios? If ct_state can become -trk, I feel, there will be chances of loop getting created between tables 244 and 242.

It should be never be -trk when submitted to netfilter, it is a bug in ovs that makes it -trk  for packets created in the pipeline like the ping responder which is now handled with the fix. Once Ovs fixes the bug the workaround can be removed. Since the other packets works fine with ovs, I don't see a reason for it to be -trk. Also we are bypassing netfilter for non supported packets now, so that shouldn't be a problem either.

Patch-set: 2