NETVIRT-1235 : Prefix-To-Interface ds not cleared .
We have 2 primary adj  stored in VpnInterfaceOp DS. When VM is deleted , 2 FIB entries are removed.
In cleanupOpDataForFib() , when we remove first primary adj from VpnInterfaceOp DS , it triggers update in VpnInterfaceOp listener.
We ignore update of VpnInterfaceOp. When 2^nd^ / final primary adj is removed , we delete the VpnInterfaceOp entry.
This leads to removal of prefix-to-interface for 2nd primary adj. First primary adj , always remains stale.

Change-Id: I8fddd9a5d1a7425533ca1c45f81385f60bfa0d4f
Signed-off-by: eaksahu <a.k.sahu@ericsson.com>