Update patch set 7
Patch Set 7:

(1 comment)

> Hi Sridhar,
 > 
 > For Subnet DHCP IPs are programmed in FIB entry is not because of
 > this PS (code changes). Have observed that behavior without this
 > code changes.
 > 
 > Request your input for addressing the reported issue.
 > 
 > I would think for addressing DHCP IPs in FIB issue with different
 > bug? Please let me know if you have any concern about this.
 > 
 > Thanks in Advance,
 > Karthikeyan.

Yes Karthik, I agree that DHCP Port FIB entries are not because of this patch. But since the bug/patch is trying to address a use-case where we want to avoid IPv4 FIB entries when the subnet is not associated to the router, I feel its relevant to the same bug.

But if you want to address it in a separate patch, I'm fine. Please don't close the bug (Netvirt#1171) but add the new patch-set to the same bug.

Patch-set: 7
Label: Verified=0