Update patch set 2
Patch Set 2: Code-Review-1

This fix need to be altered so that we avoid 'backpull' from the VRFEngine.

Basically, we cannot try and pull subnetmap from within the VRFEngine (ie., VRFEntryListener) as that is bound to fail when we increase the rate of orchestration of router-interfaces.

So please let us find an alternate way to recover the ip-class information instead of backpulling from subnetmap within the VrfEntryListener.

Patch-set: 2
Label: Code-Review=-1