Update patch set 5
Patch Set 5: Code-Review-1

This change is not required for replay-based-upgrade as the code to populate extra route adjacencies on upgrade is already in place.
This review might be required for IpSerialization. So holiding off merge for now.

Patch-set: 5
Reviewer: Gerrit User 5871 <5871@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1