Router interface routes should not be exported
Router-interface-ip addresses are not exposed as BGP routes,
as a result we should not be handling such routes for Importing VPNs.

Please note that we do not create an operational vpn-interface for a
router-interface, however we do create a fib-entry for a
router-interface which in turn used by VrfEngine to program the
L3_GW_MAC_TABLE (table 19) and the PING RESPONDER rules in FIB-TABLE
(table 21).

Change-Id: I2f442868db3513ee196b14ddc5cf9ba2577641cc
Signed-off-by: Deepthi V V <deepthi.v.v@ericsson.com>