Helper patch for handling external router changes
While implementing IPv6 External Connectivity (FLAT/VLAN based provider networks),
we would need a listener to handle changes to the external routers YANG model.
It is seen that the current implementation of ExternalRoutersListener is not
properly modularized. Making any further changes in that class could introduce
regressions for Controller, Conntrack based SNAT and BGPVPN use-cases. So, this
patch adds a new listener SnatExternalRoutersListener for the Routers YANG model,
but only one of the two listeners would be active at a time as we register the
listener depending on the NatMode that is configured in the deployment.

This patch mainly re-organizes the code and does not introduce any functional
changes. Its mainly a helper patch and the necessary IPv6 pipeline changes
would be done in a separate patch.

Change-Id: I424de2494abb1140b1f4d702d483498f9f072363
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>
(cherry picked from commit 1d8d695790ffde338aa4c424cb0bc8a5db122f4b)