NETVIRT-1278: Don't write mac-in-use to config DS
When a node is discovered for which there is no entry
in the config datastore we copy it over so that we can
than make modifications to it. A new field, mac-in-use,
was added and it is operational only so when attempting
to write the bridge to config the whole thing fails.

Fix is simple, just don't write the termination points
to the config datastore at all.

Change-Id: I0592d34fb0500e2e8001be617af8b11a5eb9d6c9
Signed-off-by: Josh <jhershbe@redhat.com>