When number of RDs are lesser than number of nexthops
In checkAlarmExtraRoutes, when number of RDs are lesser
than number of nexthops, ECMP code was not getting
triggered.

Hence, when number of RDs are lesser
than number of nexthops, then just raise the alarm.

Previously, in routesError there are a few route raised
in alarm, so they are not to be used, but now, we are not
maintaining any error routes.

Change-Id: If0a1ef22d157b51aa74cad2eb3c7af27b70933dc
Signed-off-by: eswanit <swati.udhavrao.niture@ericsson.com>