Supply correct tunnel type to get tunnel names
A recent change in ITM RPC has changed the API to retrieve the tunnel
interface name. It now fetches the interface name based on the tunnel-type
provided by the called insteady of hardcoding the tunnel type to
MplsOverGRE.
L3VPN was always sending tunnel type as VxLAN when calling this RPC.
The BGP routes were not being translated to flows.
The fix supplies the correct tunnel-type to the RPC based on the encap
type of fibEntry

Change-Id: Id5159ce8bc1ff686defe9121c8b4b7edfd809bbf
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>