unblock broken build by replacing some but not all Xtend from aclservice
This rewrites some of the Xtend with Java.  It only replaces those
*Builder classes, but does not (yet?) replace the expected objects with
some TBD not Xtend based approach.

As it is something related to Java generics in those Builder classes
which seems to (suddenly) somehow be causing the build failure, this is
sufficient, for now, to unblock broken build.

JIRA: NETVIRT-1290
Change-Id: Id76d24f1475ed3d3eeac622de8932fc8a63f8ee6
Signed-off-by: Michael Vorburger <vorburger@redhat.com>