Update patch set 1
Patch Set 1:

> > This seems rather heavy-handed, couldn’t we just @Ignore them?

heavy handed, with a heavy heart - but it seemed the right thing to do.. but read on.

 > The
 > > difference being that @Ignored code still gets refactored when
 > > necessary...
 > 
 > Although given where the failure is occurring this might not be
 > possible.

exactly; the build failure was in when the xtend-maven-plugin transpiled some of the *.xtend to *.java - so @Ignore on @Test won't fix that.

but I'm guessing that you'll like https://git.opendaylight.org/gerrit/#/c/72571/ ? While writing this, it occurs to me that perhaps I should just "squash" these two now actually?

Patch-set: 1