Update patch set 3
Patch Set 3: Code-Review-1

> Can we use the @Deprecated API syncReadOptionalAndTreatReadFailedExceptionAsAbsentOptional()

no.. basically, the rule is (should be) to *NEVER* ever use @Deprecated APIs in new code you write... ;-)  Because anything @deprecated should explain why it is so, and what to do instead.  (I know we have things which are @Deprecated with no clear alternative; in my personal opinion, those are wrong to be deprecated.)

 > instead of syncReadOptional(). If the later is used, then an
 > explicit try-catch need to be done. If we have .present() check
 > after read operation, it's better to use syncReadOptionalAndTreatReadFailedExceptionAsAbsentOptional()
 > which return Optional.absent() during ReadFailedException

No, that's bad; because it "cheats" and lets code continue with wrong values; as per the JavaDoc on that method.  We have to deal with exceptions.  Often the right thing to do is to simply let them propagate to the initially caller who will with them - e.g. RPCs entry points, or jobcoordinator jobs that will fail, etc.

Patch-set: 3
Label: Code-Review=-1