NETVIRT-1302 SNAT Connectivity from non-FIP instances to an external IP fails when using VLAN setup (conntrack)
When the NAPT switch is hosted on a node without the vlan footprint,
SNAT from the instace in the vlan on other nodes are failing.

We are now creating a vlan footprint in the NAPT switch by adding a
pseudo port in the elan-dpnlist for all the vlan added to the router
when the router g/w is set.

Change-Id: I55df1109a5415aa425647539d4319fce68899938
Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>