Update patch set 10
Patch Set 10:

> Build Failed
 > 
 > https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-fluorine/458/
 > : FAILURE (skipped)
 > 
 > The job has timed out after 360 minutes. Examine output, retry, or
 > consider increasing timeout value for this job type/stream. (
 > https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-fluorine/458/
 > )
 > 
 > Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/netvirt-patch-test-current-fluorine/458

The tests passed up until the job aborted - because of a timeout. The bring up took 3 hours and the tests were running for 3 hours so it hit a 6 hour job timeout. logs are not kept in this case so it is hard to diagnose, but I suspect package downloads were taking a while. I have seen that before.

Patch-set: 10