Upgrade: wait for interface states for adjacencies
This issue was discovered in upgrade CSIT. The issue
is that usually L3 configuration comes in way after
tunnel interface states are present in the operational
database. However, in the case of upgrade this order
can be reversed. The following fix utilizes
DataTreeEventRegistrar to wait for the specific interface
states required to program the Vrfs. Note that this is
only done during upgrade.

JIRA: NETVIRT-1312

Change-Id: I3b3ba12bbb3453a83dd4969ae6d2487d2fe36dcb
Signed-off-by: Josh <jhershbe@redhat.com>