<WIP>IPv6 CVR North-South Support
This patch programs the pipeline flows on the NAPT and non-NAPT
switch to support North-South traffic. Traffic from non-NAPT
Switch would be forwarded to the NAPT Switch from where its
sent to the external network.

In order to support the IPv6 N/S end-to-end use-case, GWMACResolution
has to happen and the necessary support would be added via a
separate patch.

Change-Id: I2fa0f57af9b063d6bd87c385ab7b1b4d1add5186
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>
Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>