Update patch set 2
Patch Set 2: Code-Review-1

(15 comments)

First of all, thanks for tackling this! Some minor feedback.  Also a more general point that I probably would have directly inlined the metrics int the respective classes, instead of the indirection with those extra classes that you turned the former counter enums into - that only really makes sense, IMHO, if several classes need to increment the same metric? I can see how that is the case in some cases here, but I don't think it is everywhere here?  But up to you; I'll +1 it even if that is left as it is here, after the other minor feedback is addresses.

Patch-set: 2
Label: Code-Review=-1