LPort tag issue with controller-snat
Desciption: When packet is sent for SNAT translation
from Non-Napt Switch, the lport tag is overwritten to
tunnel lport tag(0->36 flow). During SNAT
translation, when we query for interface name from
if-indexes-interface-map DS for this tunnel lport tag
which is not available(with itm direct tunnel changes)
resulting in Exception. Changes are done to query the
dpn-id and in_port from the packet for Non-Napt Cases.

Change-Id: I9f9b6117b4acfcd3bb33adbfa86e2c25e1218976
Signed-off-by: xcheara <chetan.arakere@altencalsoftlabs.com>