Update patch set 14
Patch Set 14:

Adding Guice Rules to tests in the vpn-manager module requires considerable effort and time, none of the tests in the module have been refactored to use the new apporach of writing JUnits using Guice.
I think it would be better to take this as a separate initiative.

I have mocked required references to remove the @Ignore.

Patch-set: 14
Label: Verified=0