NETVIRT-1054 : FATAL:fib-entries container missing
If fib:show command is executed the very first time
after karaf bringup, when no routes have yet been
populated, the container will not be created. A GET
will then throw a 404.
In this case, printing a fatal messge to console
is not required. The commmand will only print an
empty table.

Change-Id: If4555c62578960ae03627832cfdc6275560fc53b
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>