Update patch set 5
Patch Set 5:

> > ASwin, tempest job still isn't looking good. can you look?
 > 
 > This is to fix a randomly occuring NPE the issue. The current
 > failures seems to be related to [1]. I have reopened the bug.
 > 
 > [1]https://jira.opendaylight.org/browse/NETVIRT-1270

NETVIRT-1270 is a filtered exception - the reason it popped up recently though, is the "ConflictingModificationAppliedException: Node children was modified by other transaction". That is now "ConflictingModificationAppliedException: Node was created by other transaction". The string doesn't match now. We can modify the whitelist to not be so specific on the modified vs created part. Let me know if that makes sense so the csit passses or if you would prefer to fix the bug. Also +1/+2 if you want to merge this patch.

Patch-set: 5