Allow MDSALManager exceptions
The exception-handling story in MDSALManager is currently
underwhelming: we basically swallow errors, or encapsulate them in
RuntimeException. To allow this to be changed, NetVirt has to allow
for exceptions coming from MDSALManager; this patch implements this.

Once MDSALManager’s exceptions have been defined, this will be
revisited to clean everything up and use specific exceptions.

JIRA: NETVIRT-1318
Change-Id: I66ad8c2643a03d64ccb4abed94628498636f9152
Signed-off-by: Stephen Kitt <skitt@redhat.com>