Prefix to interface d.s not cleared
Issue:
======
Stale flows not removed for many data stores .
Stale entries were not removed for prefix-to-interface d.s
because of which dependent vpn related d.s were not cleared.

Analysis:
=========
Prefix-to-interface was wrongly getting deleted from
removePrefixToInterfaceAdj() in updtae() in VpnInterfaceManager.java
because of which in cleanUpOpDataForFib() ,prefixinfo was null and the
respective calls to delete prefix-to-interface and vpn-interace-op-data
was getting missed.
         Removed the call removePrefixToInterfaceAdj().

	 Issue: NETVIRT-1223

Change-Id: Id892b033014606e6c451cb2c5644d839ec2da5ac
Signed-off-by: nithi thomas <nithi.t@altencalsoftlabs.com>