ECMP routes are not clearing in dualstack network
Problem Description:
====================
ECMP routes are not cleaned up properly in dual
stack network use cases.
The following neutron command is not working
properly for IPv6 address family in extra route
feature.

neutron router-update R1 --no-routes

Solution:
=========
FIBManager is not properly handled for both IPv4 and
IPv6 address family in case of dual stack network.
It is always added IP prefix length value is /32 instead of
validating whether it is IPv4 or IPv6 prefix. As a result
querying prefix-to-interface data store is returning null
value for valid IPv6 prefix. This is the root cause of the
problem for clearing all IPv6 extra routes.

As part of this fix, have handled properly for both IPv4 and IPv6
address family.

Issue: NETVIRT-1355

Change-Id: Iadb69e0c259d125c20c825edbdc7716495eb077a
Signed-off-by: Karthikeyan Krishnan <karthikeyangceb007@gmail.com>