Update patch set 2
Patch Set 2:

> > Also it looks like this is not cherry-picked (but an independent
 > > patch). Ideally you should cherry-pick the patch and if there are
 > > any manual changes include them in the commit message as
 > > "Conflicts".
 > 
 > Cherry-pick was failing for me. I missed editing the commit
 > message.
 > I have edited the commit message, it got generated as a new gerrit
 > link. The below is the one.
 > 
 > https://git.opendaylight.org/gerrit/#/c/74180/

Sure if the gui cherry-pick fails, then you manually cherry-pick in your local branch. In that case, after you do the git commit -s and a ChangeId: hash is added to your patch, do a commit --amend and then change the hash to the ChangeId hash in the master patch. Then push your patch. You can do that last amend step now and push again. The ChangeId is what links the patches. Also note - you can not edit the commit-message via this gerrit gui and have it update the cherry-pick links. Your edit will be reflected in the commit-message but the link doesn't happen. The link only seems to be created when you push a patch.

Patch-set: 2