Avoid ofp topology discovery
openflowplugin is packaging lldp based topology discovery
along with "odl-openflowplugin-flow-services. This feature sends
periodic lldp messages on all the links available, to discover
the topology. Netvirt does not need this lldp feature, but is coming
by default with the feature we were including.

This patch removes the feature which was previously used, and
use only the appropriate openflowplugin features needed for netvirt
functionality.

JIRA: OPNFLWPLUG-1024

Change-Id: Ibdd78e492c77fb19ae1965bb6c92a6b4a84dd695
Signed-off-by: Faseela K <faseela.k@ericsson.com>