Update patch set 2
Patch Set 2: Code-Review+1

This LGTM so +1, but I would love Josh and Faseela (and perhaps others; have added Aswin original author of https://git.opendaylight.org/gerrit/#/c/74011/ where this comes from) to +1 this as well...  

If we all to agree and merge this, then we should perhaps explain this further in future at least genius and netvirt weekly meetings, and send out some email elaborating.  

Strictly speaking I guess this does slighty change the behaviour, as in the case it's already present, currently it's sync while this makes it async.. but it COULD be async anyway today (if it's not present), so this makes it more consistent, which is probably a Good Thing?

Patch-set: 2
Label: Code-Review=+1