Update patch set 6
Patch Set 6:

(2 comments)

> keeps failing a few tests

I’ve spent some (too much) time to review this change in detail, to see if I could spot some typo like a mixed up Configuration VS Operational or mixed up arguments or something like that.. 

It seems mostly fine to me.. but did find 1 change to a delete in VrfEntryListener which looks suspicious - not sure if that could be causing these CSIT failures?  Stephen, shall we change that back, or was that intentional / fine / not a problem?

Otherwise, if (when) we have an otherwise 100% stable CSIT, Sam can you re-run this, and if it still fails, help dig? Are there any (new) errors in the karaf.log which could be related to this change? For example if there were any complaints from the TransactionAdapter (IllegalArgumentException re. "Invalid datastore %s used instead of %s"), that would be very interesting.

Patch-set: 6
Label: Verified=0