Update patch set 12
Patch Set 12:

Hi Sam/Tom/Michael,

There are CSIT failures due to new Exceptions (due to ModifiedNodeDoesNotExistException) in this patch.
The code tries to delete adjacency from VpnInterface object.
When VpnInterface (parent) itself is deleted, there is no need to delete adjacency(child) as it is taken care.
Hence ModifiedNodeDoesNotExistException is an expected exception when VpnInterface is deleted before deleting adjacency.

So, exception is not logged in VpnUtil.removeMipAdjacency() method but it is getting logged as WARN in controller project's ShardDataTree.java/LocalThreePhaseCommitCohort.java

I can think of two options to make CSIT pass:
1. Remove WARN/ERROR log messages in controller project's ShardDataTree.java/LocalThreePhaseCommitCohort.java files.
2. Exclude WARN log messages being considered as CSIT failures.
 
Request your inputs on this issue, please.

Thanks in Advance,
Som

Patch-set: 12