Update patch set 1
Patch Set 1:

Initial save of work before I head out for vacation. Refactored NatInterfaceStateListener - extracted methods to make a NatSouthboundEventHandler that can be invoked from the NatArpNotificationHandler. TBD: tie those together and implement housekeeping so that (a) arp flows not continuously responded to and (b) when the VIP moves

Patch-set: 1