OptimisticLockFailedException fix for IPv6 Serv
Issue:
=======
OptimisticLockFailedException is throwing for
when IPv6 host interface service binding and
unbinding scenario. In cluster environment
this issue was observed.

Solution:
=========
As pert of this fix, have introduced JC for
IPv6 service binding and unbinding for all
the basic add(), update(), remove() methods.
This kind of JC handling has implemented already
in all the southbound events. So the same kind
of handling is implemented for all the
northbound events.

Issue: NETVIRT-1409

Change-Id: Ice777d57913cda8924ec8422eb02bbbfdd194168
Signed-off-by: Karthikeyan Krishnan <karthikeyangceb007@gmail.com>