neutron-vpn-portip-port-data d.s not cleared
Issue:
======
Stale entry not removed for neutron-vpn-portip-port-data .

Analysis:
=========
neutron-vpn-portip-port-data d.s int getting cleared when router
is attached to BGP-VPN and then all the configs are deleted.
When router is attached to BGP-VPN dhcp interfaces are also getting added
to neutron-vpn-portip-port-data d.s which shhould not happen. This doesnot
happen when subnets are attached to router.
     When router is attached to BGP-VPN updateVpnInterface() is called
where the check of blocking dhcp interfces was missed out. Added the checks.

Issue: NETVIRT-1406

Change-Id: Idaf6f43525655c4043f5cee7b32263cd9858e31c
Signed-off-by: nithi thomas <nithi.t@altencalsoftlabs.com>