VPNInstance IP AF update is not working
Issue:
======
VPNInstance data is not updating correctly for dual stack
network use cases. Based on the added subnet address family
it should update properly. As a result of improper address
family update bgp routes to external communication was
not working properly.

Solution:
============
As part of this fix have addressed the proper way of
updating the VPNInstance data store with IP address family
based on the attached subnets to the router.

Issue: NETVIRT-1410

Change-Id: I9dda8beeb74aaa273db800964b5e741478db36a3
Signed-off-by: Karthikeyan Krishnan <karthikeyangceb007@gmail.com>