Stale neutronrouterdpn Oper DS cleanup
Description:
On deletion of router, its been observed that there exists an invalid
empty entry for router in neutron-router-dpn DS. This DS will be
populated whenever an VM created/deleted on a DPN part of the router
by NAT. On deletion of all VM on a given router, the empty router
entry left over and even not clear on router deletion.

Changes done to clear the router entry on from this DS on
router deletion.

Issue : NETVIRT-1406

Change-Id: I0db96497f1da92e38268a966066d46b1bfee9e6c
Signed-off-by: xcheara <chetan.arakere@altencalsoftlabs.com>
Signed-off-by: shaik <shaik.b@altencalsoftlabs.com>