Local Bucket Entry deleted
Verify ECMP traffic distribution with weighted buckets -
Controller Restart with CSS in standalone Mode
========================================================

After Controller Restart the Local Bucket Entry was getting deleted.
This was happening bacause of multiple simultanious event was causing
ConflictModifiactionException while updating the group entry. Now
we are using DJC to submit a job using SAME key for adding group and
flows, so that it would go to the samed jobqueue and writes get
executed serially, first the froups and then the flows), instead of
parallely and no comflict in writing by diffent thread to the same
node.

Change-Id: I23ad23d1db830213feb992897ee01de6148b156a
Signed-off-by: eceghkl <manu.b@ericsson.com>