Remove GPE requirement for SFC classifier
GPE is the canonical way to encapsulate NSH in VXLAN. But it can
be done without at the expense of the bandwidth wasted with a
useless Ethernet header which is a useful option for pre-existing
VXLAN meshes.

For the time being, SFC is not using GPE extension. And the plan
is to not make it a hard requirement but an option in the (near)
future. When the time comes, support both mechanisms.

Change-Id: I120f8a010e3b6e21a4371eef7e114ac53745f190
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>