NETVIRT-1439 : Delete L3VPN cleanup failure.
After running clean script seeing stray fib and
vpn configs

A recent fix to handle cleanup of external PNF flows had broken
footprint cleanup of all other types of FIB entries. This fix made use of
the parentVpnRd attribute in VrfEntry to check for the footprint of
external subnet where the PNFs are discovered (Since, the
footprint/interfaces for that external subnet is maintained in the VPN of
its external network). The parentVpnRf points to this external network.

The above mentioned fix broke cleanup of existing FIBs since no other
non-PNF fibs will hve their parentVPnRd set.

SO, this fix considers the primaryRd if parentVpnRd is not set.

Change-Id: I3a6f216e9f9cfa6bc6fe1dc8fa7e95b958ebd1b7
Signed-off-by: Kiran N Upadhyaya <kiran.n.upadhyaya@ericsson.com>
Signed-off-by: Achuth Maniyedath <achuth.maniyedath@gmail.com>