Remove DS operation part of Yang Consumer Thread
Description:
1. Changes are done to remove move the logic of
performing DS read/write and install
installation/removal to perform in the worker
thread instead of performing the same in yang
consumer thread.

2. In the private BGPVPN association case, if the
association is created first and the router-gateway-set
is done later, it is noticed that the default NAT
flow is instaled with external VPN metadata rather then
private BGPVPN metadat. Changes are done to correct it

JIRA : NETVIRT-1392

Change-Id: Ic296b2d3c1c22350518c5008ceb58df1ef17a76c
Signed-off-by: shaik <shaik.b@altencalsoftlabs.com>