Update patch set 1
Patch Set 1:

> Build Started https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-neon/221/

this one is likely to timeout also... this started 3 hours ago, but the downstream job didn't start until an hour ago. But since it is started it will finish, just need to use the link below to find it. It should pass like the previous, so we will wait and see then merge this patch.
 
https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-0cmb-1ctl-2cmp-openstack-queens-gate-stateful-neon/264/

Patch-set: 1