Stale flows in ACL tables 216/246
Issue:
------
Stale flows with VM IP's are observed in ACL tables 216/246 when a VM
is deleted which was having SG with --remote-group rules.

Solution:
---------
The issue is due to race-conditions.
Processing VM deletion is handled via DJC where-as deletion of
AclInterface from cache is not. This was resulting in stale flows.
Now, DJC with same key (ACL-ID) is used even for deletion of
AclInterface from cache which will sequence the operations.

JIRA: NETVIRT-1469

Change-Id: I315831b62a2ac5954fa4177a755c828a1f43d476
Signed-off-by: Somashekar Byrappa <somashekar.b@altencalsoftlabs.com>