NETVIRT-1410 IPv4 and IPv6 routes were not adveritised in DC-GW with dual stack scenario
On a Router/Network association, the vpn-op update to configure IpFamily
was being done after the creation of Vpn Interfaces. THis was leading to
the Networks being written to bgp before the vrf, causing failure in
advertisement of the route to DC-Gw.
This fix swaps the order of IpFamily update and VPN Interface creation.

Change-Id: I52ebf0b5abd9869dfe1c08b2f2511d0ec8468a02
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>