ECMP - MIP/learnt IPs pointing to groups
Issue:
MIP/learnt IPs (both local and remote) programmed in FIB
(table-21) were pointing to load balancing groups rather
than redirecting to table-220. Only extra routes
(in table 21) need to be pointed to the group.

Fix:
Adding a check to enter setupLoadBalancingNextHop if
it is a static route

Also, the logic for removeAdjacencyFromInternalVpn method
is changed for the deletion of extra-route and nexthop.

Change-Id: Ie2aa43fb3180c032a609cba4fa39ac20f3a5b003
Signed-off-by: eswanit <swati.udhavrao.niture@ericsson.com>