NETVIRT-1517 : Stale MIP FIB/Flow entires present upon deletion of VRRP master
Issue:

This issue is seen when ports are pre-created and VM is booted on that.
In such condition , when VM is deleted , only south bound Interface remove
is received.
When VM which is a VRRP master(holding MIP-IP) was deleted , MIP-IP adj
was not deleted from the config-VpnInterface. Thus leading to issues like
duplicate FIB entry , when VM is booted back.

Fix:
MIP-IP adj delete is handled appropriately when VM is deleted.

Change-Id: I304bf57e7247bb1942d14e1b48933e5f91aa8d1c
Signed-off-by: Achuth <achuth.maniyedath@gmail.com>