NETVIRT-1520: AlivenessMonitor cache issue
Issue:
Say VM with MIP-IP is on DPN connected to Contorller-1, when MIP is
learnt , it populates the cache alivenessCache. When VM migrates to
another DPN, which is connected to say Controller-2, the same
alivenessCache will not have data w.r.t the monitoring session
created by Controller-1.
Hence incase of cleanup , it causes issues in stopping ArpMonitoring.
The cache is simple local cache which doesnt work incase of cluster.

Fix:
Better to read InterfaceMonitorMap directly from datastore so that
monitorId can be retrieved even if MIP moves across DPNs connected to
various controllers in cluster.

Change-Id: I1e4102a7c57bff8376d3d05be8975a6c1ad74233
Signed-off-by: Anil Kumar Gujele <anilkumar.g@altencalsoftlabs.com>